Moved form related components from visits folder into its own folder
